Integerated auth0 and protected all the routes

+++ pages/api/deleteSnippet.js
@@ -1,10 +1,17 @@
+import { getSession, withApiAuthRequired } from "@auth0/nextjs-auth0"
 import { deleteSnippet } from "../../utils/Fauna"
 
-export default async function handler(req, res) {
+export default withApiAuthRequired(async function handler(req, res) {
   if (req.method !== 'DELETE') {
     return res.status(405).json({ msg: 'Method not allowed.' })
   }
+  const session = getSession(req, res)
+  const userId = session.user.sub
   const { id } = req.body
+  const existingRecord = await getSnippetById(id)
+  if (!existingRecord || existingRecord.data.userId !== userId) {
+    return res.state(404).json({ msg: 'Record not found.' })
+  }
   try {
     const deleted = await deleteSnippet(id)
     return res.status(200).json(deleted)
@@ -13,4 +20,4 @@ export default async function handler(req, res) {
     console.error(err)
     res.status(500).json({ msg: 'Something went wrong.' })
   }
-}
+})
